server/requirements.txt
Flask
psycopg2
flask_sqlalchemy
Flask-Migrate
flask_cors
Authlib
flask-jwt-extended[asymmetric_crypto]
argon2-cffi

# Api docs
apispec
apispec-webframeworks
flasgger
flask-marshmallow
marshmallow
marshmallow-sqlalchemy